[BEHAVIORAL] Add Windows-compatible process liveness check for checkpoint recovery
isProcessAlive used Unix-only syscall.Signal(0) which doesn't work on
Windows. Now split into platform-specific files:

- process_unix.go: uses syscall.Signal(0) (existing behavior)
- process_windows.go: uses windows.OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION)
- Removed syscall import and inline function from recovery.go
- 3 new tests: current process alive, dead PID, negative PID
